# openapi.yaml
openapi: 3.0.3
info:
  title: Control of Moedinhas API
  description: API para gerenciamento de domicílios, usuários, contas, categorias, transações e divisão de despesas.
  version: "1.0.0"
servers:
  - url: http://localhost:8000
paths:
...
  /categories/:
    post:
      summary: Criar categoria
      tags: [Categories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        '201':
          description: Categoria criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryCreatedResponse'
    get:
      summary: Listar categorias
      tags: [Categories]
      parameters:
        - name: household_id
          in: query
          schema:
            type: string
          required: false
      responses:
        '200':
          description: Lista de categorias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
  /categories/{category_id}:
    get:
      summary: Obter categoria
      tags: [Categories]
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Categoria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
  /accounts/:
    post:
      summary: Criar conta
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreate'
      responses:
        '201':
          description: Conta criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountCreatedResponse'
    get:
      summary: Listar contas
      tags: [Accounts]
      parameters:
        - name: household_id
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Lista de contas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
  /accounts/{account_id}:
    get:
      summary: Obter conta
      tags: [Accounts]
      parameters:
        - name: account_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Conta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
  /transactions/:
    post:
      summary: Criar transação
      tags: [Transactions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCreate'
      responses:
        '201':
          description: Transação criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionCreatedResponse'
    get:
      summary: Listar transações
      tags: [Transactions]
      parameters:
        - name: household_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de transações
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionDB'
  /transactions/entries/month/{year}/{month}:
    get:
      summary: Listar lançamentos por mês
      tags: [Transactions]
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: integer
        - name: month
          in: path
          required: true
          schema:
            type: integer
        - name: household_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lançamentos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionEntry'
  /transactions/entries/card/{account_id}/{year}/{month}:
    get:
      summary: Extrato de cartão por conta e mês
      tags: [Transactions]
      parameters:
        - name: account_id
          in: path
          required: true
          schema:
            type: string
        - name: year
          in: path
          required: true
          schema:
            type: integer
        - name: month
          in: path
          required: true
          schema:
            type: integer
        - name: household_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Extrato do cartão
          content:
            application/json:
              schema:
                type: object
                properties:
                  competence_month:
                    type: string
                  total:
                    type: number
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/TransactionEntry'
  /transactions/entries/{entry_id}/pay:
    patch:
      summary: Marcar lançamento como pago
      tags: [Transactions]
      parameters:
        - name: entry_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lançamento marcado como pago
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericMessage'
...
components:
  schemas:
    GenericMessage:
      type: object
      properties:
        message:
          type: string
...
    CategoryCreate:
      type: object
      required: [name, type, household_id]
      properties:
        name:
          type: string
        type:
          type: string
          description: "income | expense"
        household_id:
          type: string
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        household_id:
          type: string
    CategoryCreatedResponse:
      type: object
      properties:
        message:
          type: string
        category_id:
          type: string
        name:
          type: string
        type:
          type: string
        household_id:
          type: string
    AccountCreate:
      type: object
      required: [name, type, initial_balance, household_id]
      properties:
        name:
          type: string
        type:
          type: string
          description: "checking | savings | credit_card"
        initial_balance:
          type: number
        closing_day:
          type: integer
          nullable: true
        due_day:
          type: integer
          nullable: true
        household_id:
          type: string
    Account:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        initial_balance:
          type: number
        current_balance:
          type: number
        household_id:
          type: string
        closing_day:
          type: integer
        due_day:
          type: integer
    AccountCreatedResponse:
      type: object
      properties:
        message:
          type: string
        account_id:
          type: string
        name:
          type: string
        type:
          type: string
        initial_balance:
          type: number
        household_id:
          type: string
    PaymentMethod:
      type: object
      properties:
        type:
          type: string
          enum: [cash, debit, credit_card, pix]
        closing_day:
          type: integer
          nullable: true
        due_day:
          type: integer
          nullable: true
    SplitUser:
      type: object
      properties:
        user_id:
          type: string
        amount:
          type: number
          nullable: true
        percentage:
          type: number
          nullable: true
        paid:
          type: boolean
    TransactionCreate:
      type: object
      required: [description, amount, flow_type, transaction_type, payment_method, date]
      properties:
        description:
          type: string
        amount:
          type: number
        flow_type:
          type: string
          enum: [income, expense]
        transaction_type:
          type: string
          enum: [single, installment, recurring]
        payment_method:
          $ref: '#/components/schemas/PaymentMethod'
        date:
          type: string
          format: date
        installments:
          type: integer
        total_amount:
          type: number
        account_id:
          type: string
        category_id:
          type: string
        household_id:
          type: string
        split_type:
          type: string
        split_users:
          type: array
          items:
            $ref: '#/components/schemas/SplitUser'
        paid_by_user_id:
          type: string
    TransactionDB:
      type: object
      properties:
        id:
          type: string
        description:
          type: string
        amount:
          type: number
        flow_type:
          type: string
        transaction_type:
          type: string
        payment_method:
          $ref: '#/components/schemas/PaymentMethod'
        date:
          type: string
          format: date-time
        account_id:
          type: string
        category_id:
          type: string
        household_id:
          type: string
    TransactionCreatedResponse:
      type: object
      properties:
        message:
          type: string
        transaction_id:
          type: string
        entries_created:
          type: integer
    TransactionEntry:
      type: object
      properties:
        id:
          type: string
        transaction_id:
          type: string
        description:
          type: string
        amount:
          type: number
        competence_month:
          type: string
        due_date:
          type: string
          format: date
        status:
          type: string
        payment_method:
          type: string
        account_id:
          type: string
        category_id:
          type: string
        household_id:
          type: string
...
security: []